+++ src/soccergoals/__main__.py
@@ -1,3 +1,4 @@
 from soccergoals.main import main
 
-main()
+if __name__ == "__main__":
+    main()